Replace the last occurrence of 'min' in the file name instead of the first
Previously, if you used any of the words in this list
http://www.morewords.com/contains/min/ to name your bundle (say, for
example, `minimal.min.js`) then this plugin would strip the first min
resulting in a file named `imal.min.js` which makes no sense.

This change switches the `min` replacement to look for the last
occurrence of min instead of the first, since that is the more common
naming convention for minified files.